reorganise the hierachy of packages
