+++ document/document_test.go
@@ -6,10 +6,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	_ "github.com/ucan-wg/go-did-it/methods/did-key"
-	"github.com/ucan-wg/go-did-it/verifications/ed25519"
-	"github.com/ucan-wg/go-did-it/verifications/jsonwebkey"
-	"github.com/ucan-wg/go-did-it/verifications/x25519"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/ed25519"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/jsonwebkey"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/x25519"
+	_ "github.com/ucan-wg/go-did-it/verifiers/did-key"
 )
 
 func TestRoundTrip(t *testing.T) {
