+++ verifiers/_methods/ed25519/VerificationKey2018.go
