+++ verifiers/_methods/ed25519/VerificationKey2020.go
