+++ verifiers/_methods/ed25519/VerificationKey2020_test.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/ucan-wg/go-did-it"
 	"github.com/ucan-wg/go-did-it/crypto/ed25519"
-	_ "github.com/ucan-wg/go-did-it/methods/did-key"
-	"github.com/ucan-wg/go-did-it/verifications/ed25519"
+	ed25519vm "github.com/ucan-wg/go-did-it/verifiers/_methods/ed25519"
+	_ "github.com/ucan-wg/go-did-it/verifiers/did-key"
 )
 
 func TestJsonRoundTrip2020(t *testing.T) {
