+++ verifiers/_methods/json.go
@@ -1,16 +1,16 @@
-package verifications
+package methods
 
 import (
 	"encoding/json"
 	"fmt"
 
 	"github.com/ucan-wg/go-did-it"
-	"github.com/ucan-wg/go-did-it/verifications/ed25519"
-	"github.com/ucan-wg/go-did-it/verifications/jsonwebkey"
-	"github.com/ucan-wg/go-did-it/verifications/multikey"
-	p256vm "github.com/ucan-wg/go-did-it/verifications/p256"
-	secp256k1vm "github.com/ucan-wg/go-did-it/verifications/secp256k1"
-	"github.com/ucan-wg/go-did-it/verifications/x25519"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/ed25519"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/jsonwebkey"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/multikey"
+	p256vm "github.com/ucan-wg/go-did-it/verifiers/_methods/p256"
+	secp256k1vm "github.com/ucan-wg/go-did-it/verifiers/_methods/secp256k1"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/x25519"
 )
 
 func UnmarshalJSON(data []byte) (did.VerificationMethod, error) {
