+++ verifiers/_methods/multikey/multikey_test.go
@@ -6,8 +6,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	_ "github.com/ucan-wg/go-did-it/methods/did-key"
-	"github.com/ucan-wg/go-did-it/verifications/multikey"
+	_ "github.com/ucan-wg/go-did-it/verifiers/did-key"
+
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/multikey"
 )
 
 func TestJsonRoundTrip(t *testing.T) {
