+++ verifiers/_methods/secp256k1/VerificationKey2019_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	secp256k1vm "github.com/ucan-wg/go-did-it/verifications/secp256k1"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/secp256k1"
 )
 
 func TestJsonRoundTrip(t *testing.T) {
