+++ verifiers/_methods/x25519/KeyAgreementKey2020_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	x25519vm "github.com/ucan-wg/go-did-it/verifications/x25519"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/x25519"
 )
 
 func TestJsonRoundTrip2020(t *testing.T) {
