+++ verifiers/did-key/document.go
