+++ verifiers/did-key/testvectors/vectors.go
@@ -13,10 +13,10 @@ import (
 	"github.com/ucan-wg/go-did-it/crypto/jwk"
 	"github.com/ucan-wg/go-did-it/crypto/p256"
 	"github.com/ucan-wg/go-did-it/crypto/secp256k1"
-	ed25519vm "github.com/ucan-wg/go-did-it/verifications/ed25519"
-	"github.com/ucan-wg/go-did-it/verifications/jsonwebkey"
-	p256vm "github.com/ucan-wg/go-did-it/verifications/p256"
-	secp256k1vm "github.com/ucan-wg/go-did-it/verifications/secp256k1"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/ed25519"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/jsonwebkey"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/p256"
+	"github.com/ucan-wg/go-did-it/verifiers/_methods/secp256k1"
 )
 
 // Origin: https://github.com/w3c-ccg/did-key-spec/tree/main/test-vectors
